initial commit
Some initial files I've been using for testing. dummyData1 for some numbers to play with is from the drive. plot_drafting.ipynb is a jupyter notebook so I can mess with plots in an environment I'm familiar with. And dashtest.py is the file I've been using for trying to figure out this dash thing. I think I got bootstrap to work?!?!?